pyproject.toml
[project]
name = "datarepo-rerun-viewer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "distinctipy>=1.3.4",
    "ipython>=8.31.0",
    "minio>=7.2.14",
    "psycopg>=3.2.4",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "rerun-sdk==0.22.0",
    "torch>=2.5.1",
    "uuid>=1.30",
]
